Organize Grid into its own folder and split up its internal classes.
