+++ src/Avalonia.Controls/Grid/GridSplitter.cs
